update fork to spdk-v19.10

+++ scripts/ceph/start.sh
@@ -8,7 +8,7 @@ script_dir=$(readlink -f $(dirname $0))
 
 base_dir=/var/tmp/ceph
 mon_ip=$1
-mon_dir=${base_dir}/mon.a/
+mon_dir=${base_dir}/mon.a
 pid_dir=${base_dir}/pid
 ceph_conf=${base_dir}/ceph.conf
 mnt_dir=${base_dir}/mnt
@@ -16,7 +16,7 @@ dev_backend=/dev/ceph
 image=/var/tmp/ceph_raw.img
 dev=/dev/loop200
 
-umount $dev || true
+umount ${dev}p2 || true
 losetup -d $dev_backend || true
 
 # partition osd
@@ -51,7 +51,7 @@ kpartx ${dev}
 
 # prep osds
 
-mnt_pt=${mnt_dir}/osd-device-0-data/
+mnt_pt=${mnt_dir}/osd-device-0-data
 mkdir -p ${mnt_pt}
 mkfs.xfs -f /dev/disk/by-partlabel/osd-device-0-data
 mount /dev/disk/by-partlabel/osd-device-0-data ${mnt_pt}
@@ -63,12 +63,13 @@ echo -e "\t[mon.a]" >> "$ceph_conf"
 echo -e "\tmon addr = ${mon_ip}:12046" >> "$ceph_conf"
 
 # create mon
-rm -rf ${mon_dir}/*
+rm -rf "${mon_dir:?}/"*
 mkdir -p ${mon_dir}
 mkdir -p ${pid_dir}
+rm -f /etc/ceph/ceph.client.admin.keyring
 
 ceph-authtool --create-keyring --gen-key --name=mon. ${base_dir}/keyring --cap mon 'allow *'
-ceph-authtool --gen-key --name=client.admin --set-uid=0 --cap mon 'allow *' --cap osd 'allow *' --cap mds 'allow *' --cap mgr 'allow *' ${base_dir}/keyring
+ceph-authtool --gen-key --name=client.admin --cap mon 'allow *' --cap osd 'allow *' --cap mds 'allow *' --cap mgr 'allow *' ${base_dir}/keyring
 
 monmaptool --create --clobber --add a ${mon_ip}:12046 --print ${base_dir}/monmap
 
@@ -79,6 +80,8 @@ cp ${base_dir}/keyring ${mon_dir}/keyring
 cp $ceph_conf /etc/ceph/ceph.conf
 
 cp ${base_dir}/keyring /etc/ceph/keyring
+cp ${base_dir}/keyring /etc/ceph/ceph.client.admin.keyring
+chmod a+r /etc/ceph/ceph.client.admin.keyring
 
 ceph-run sh -c "ulimit -n 16384 && ulimit -c unlimited && exec ceph-mon -c ${ceph_conf} -i a --keyring=${base_dir}/keyring --pid-file=${base_dir}/pid/root@$(hostname).pid --mon-data=${mon_dir}" || true
 
@@ -92,7 +95,7 @@ uuid=$(uuidgen)
 ceph -c ${ceph_conf} osd create ${uuid} $i
 ceph-osd -c ${ceph_conf} -i $i --mkfs --mkkey --osd-uuid ${uuid}
 ceph -c ${ceph_conf} osd crush add osd.${i} 1.0 host=$(hostname) root=default
-ceph -c ${ceph_conf} -i ${mnt_dir}/osd-device-${i}-data/keyring auth add osd.${i} osd "allow *" mon "allow profile osd" mgr "allow"
+ceph -c ${ceph_conf} -i ${mnt_dir}/osd-device-${i}-data/keyring auth add osd.${i} osd "allow *" mon "allow profile osd" mgr "allow *"
 
 # start osd
 pkill -9 ceph-osd || true
