+++ test/vhost/readonly/enabled_readonly_vm.sh
@@ -11,10 +11,10 @@ test_file_name="some_test_file"
 function error()
 {
 	echo "==========="
-	echo -e "ERROR: $@"
+	echo -e "ERROR: $*"
 	echo "==========="
 	umount "$test_folder_name"
-	rm -rf "$testdir/$test_folder_name"
+	rm -rf "${testdir:?}/${test_folder_name:?}"
 	exit 1
 }
 
@@ -31,27 +31,23 @@ if [[ ! -b "/dev/$disk_name"1"" ]]; then
 	error "Partition not found!"
 fi
 
-mkdir $testdir/$test_folder_name
-if [[ $? != 0 ]]; then
+if ! mkdir $testdir/$test_folder_name; then
 	error "Failed to create test folder $test_folder_name"
 fi
 
 echo "INFO: Mounting partition"
-mount /dev/$disk_name"1" $testdir/$test_folder_name
-if [[ $? != 0 ]]; then
+if ! mount /dev/$disk_name"1" $testdir/$test_folder_name; then
 	error "Failed to mount partition $disk_name""1"
 fi
 
 echo "INFO: Trying to create file on readonly disk"
-truncate -s "200M" $test_folder_name/$test_file_name"_on_readonly"
-if [[ $? == 0 ]]; then
+if truncate -s "200M" $test_folder_name/$test_file_name"_on_readonly"; then
 	error "Created a file on a readonly disk!"
 fi
 
 if [[ -f $test_folder_name/$test_file_name ]]; then
 	echo "INFO: Trying to delete previously created file"
-	rm $test_folder_name/$test_file_name
-	if [[ $? == 0 ]]; then
+	if rm $test_folder_name/$test_file_name; then
 		error "Deleted a file from a readonly disk!"
 	fi
 else
@@ -65,7 +61,7 @@ if ! rm $testdir/$test_file_name; then
 fi
 
 umount "$test_folder_name"
-rm -rf "$testdir/$test_folder_name"
+rm -rf "${testdir:?}/${test_folder_name:?}"
 echo "INFO: Trying to create file system on a readonly disk"
 if mkfs.ext4 -F /dev/$disk_name"1"; then
 	error "Created file system on a readonly disk!"
